refactor: FollowListCache에서 userIdList 필드의 자료형을 List<Integer> -> List<Long> 로 변경

+++ src/main/java/com/goormy/hackathon/redis/entity/FollowListCache.java
@@ -11,6 +11,6 @@ public class FollowListCache implements Serializable {
 
     @Id
     private String hashtagId;
-    private List<Integer> userIdList;
+    private List<Long> userIdList;
 
 }
