+++ src/main/java/com/goormy/hackathon/repository/FollowListRedisRepository.java
@@ -13,9 +13,9 @@ public class FollowListRedisRepository {
     private final RedisTemplate<String, Object> redisTemplate;
 
     // TODO: 수정 필요 - save 하는 쪽이 어떤 식으로 저장하느냐에 따라 호출 구현이 다를 듯
-    public List<Integer> findUserIdList(Long hashtagId) {
+    public List<Long> findUserIdList(Long hashtagId) {
         String key = "FollowList:" + hashtagId;
-        return (List<Integer>) redisTemplate.opsForValue().get(key);
+        return (List<Long>) redisTemplate.opsForValue().get(key);
     }
 
 }
